refactor: TS migration model moved to a common package
