+++ event-log/README.md
@@ -321,9 +321,34 @@ the new status.
 }
 ```
 
+- **MIGRATION_STATUS_CHANGE**
+
+Changes the status of undergoing TS migration.
+The record for which the status will be changed is defined in the `event` part of the payload.
+
+The corresponding record's status will be modified only when its current status is `SENT`.
+
+Allowed values for the `newStatus` property are: `DONE`, `NON_RECOVERABLE_FAILURE` and `RECOVERABLE_FAILURE`.
+
+**Multipart Request**
+
+`event` part:
+
+```json
+{
+  "categoryName": "MIGRATION_STATUS_CHANGE",
+  "subscriber": {
+    "url":     "http://host/path",
+    "id":      "20210302140653-8641",
+    "version": "1.22.4-10-g34454567"
+  },
+  "newStatus": "DONE"
+}
+```
+
 - **ZOMBIE_CHASING**
 
-Changes the status of a zombie event
+Changes the status of a zombie event.
 
 **Multipart Request**
