+++ event-log/src/main/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/DispatchRecovery.scala
@@ -22,10 +22,10 @@ import cats.MonadThrow
 import cats.effect.Async
 import io.renku.db.{DbClient, SqlStatement}
 import io.renku.eventlog.EventLogDB.SessionResource
-import io.renku.eventlog.subscriptions
+import io.renku.eventlog.MigrationStatus.{New, NonRecoverableFailure, Sent}
 import io.renku.eventlog.subscriptions.DispatchRecovery
 import io.renku.eventlog.subscriptions.EventsSender.SendingResult
-import io.renku.eventlog.subscriptions.tsmigrationrequest.MigrationStatus.{New, NonRecoverableFailure, Sent}
+import io.renku.eventlog._
 import io.renku.events.consumers.subscriptions.SubscriberUrl
 import io.renku.http.server.version.ServiceVersion
 import io.renku.metrics.LabeledHistogram
@@ -46,7 +46,7 @@ private class DispatchRecoveryImpl[F[_]: Async: SessionResource: Logger](queries
                                                                          now: () => Instant = () => Instant.now
 ) extends DbClient(Some(queriesExecTimes))
     with subscriptions.DispatchRecovery[F, MigrationRequestEvent]
-    with TypeSerializers {
+    with TSMigtationTypeSerializers {
   import cats.syntax.all._
   import skunk._
   import skunk.data.Completion
