+++ event-log/src/test/scala/io/renku/eventlog/TSMigrationGenerators.scala
@@ -1,0 +1,9 @@
+package io.renku.eventlog
+
+import io.renku.generators.Generators.Implicits._
+import io.renku.generators.Generators.timestampsNotInTheFuture
+import org.scalacheck.Gen
+
+object TSMigrationGenerators {
+  implicit val changeDates: Gen[ChangeDate] = timestampsNotInTheFuture.toGeneratorOf(ChangeDate)
+}
