+++ event-log/src/test/scala/io/renku/eventlog/TSMigrationModelSpec.scala
@@ -1,29 +1,9 @@
-/*
- * Copyright 2022 Swiss Data Science Center (SDSC)
- * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
- * Eidgenössische Technische Hochschule Zürich (ETHZ).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package io.renku.eventlog
 
-package io.renku.eventlog.subscriptions.tsmigrationrequest
-
-import Generators._
-import cats.syntax.all._
 import io.circe.Json
 import io.renku.data.ErrorMessage
 import io.renku.generators.Generators.Implicits._
-import io.renku.generators.Generators._
+import io.renku.generators.Generators.{nestedExceptions, nonEmptyStrings, timestamps}
 import io.renku.tinytypes.constraints.NonBlank
 import org.scalatest.matchers.should
 import org.scalatest.wordspec.AnyWordSpec
@@ -31,19 +11,8 @@ import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
 
 import java.time.Instant
 
-class MigratorSubscriptionInfoSpec extends AnyWordSpec with should.Matchers {
-
-  "show" should {
-
-    "return a String representation with the url, id and version" in {
-      val info = migratorSubscriptionInfos.generateOne
-      info.show shouldBe s"subscriber = ${info.subscriberUrl}, id = ${info.subscriberId}, version = ${info.subscriberVersion}"
-    }
-  }
-}
-
 class MigrationStatusSpec extends AnyWordSpec with ScalaCheckPropertyChecks with should.Matchers {
-  import MigrationStatus._
+  import io.renku.eventlog.MigrationStatus._
 
   "MigrationStatus" should {
