+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/DispatchRecoverySpec.scala
@@ -23,10 +23,10 @@ import EventsSender.SendingResult._
 import cats.effect.IO
 import cats.syntax.all._
 import io.renku.db.SqlStatement
-import io.renku.eventlog.InMemoryEventLogDbSpec
+import io.renku.eventlog.MigrationStatus._
+import io.renku.eventlog.TSMigrationGenerators.changeDates
+import io.renku.eventlog._
 import io.renku.eventlog.subscriptions.Generators.sendingResults
-import io.renku.eventlog.subscriptions.tsmigrationrequest.Generators.changeDates
-import io.renku.eventlog.subscriptions.tsmigrationrequest.MigrationStatus._
 import io.renku.events.consumers.subscriptions.subscriberUrls
 import io.renku.generators.CommonGraphGenerators.serviceVersions
 import io.renku.generators.Generators.Implicits._
