+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/EventFinderSpec.scala
@@ -18,12 +18,12 @@
 
 package io.renku.eventlog.subscriptions.tsmigrationrequest
 
-import Generators._
 import cats.effect.IO
 import cats.syntax.all._
 import io.renku.db.SqlStatement
-import io.renku.eventlog.InMemoryEventLogDbSpec
-import io.renku.eventlog.subscriptions.tsmigrationrequest.MigrationStatus._
+import io.renku.eventlog.MigrationStatus._
+import io.renku.eventlog.TSMigrationGenerators.changeDates
+import io.renku.eventlog.{ChangeDate, InMemoryEventLogDbSpec, MigrationStatus}
 import io.renku.events.consumers.subscriptions.{SubscriberUrl, subscriberUrls}
 import io.renku.generators.CommonGraphGenerators.serviceVersions
 import io.renku.generators.Generators.Implicits._
