+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/SubscriberTrackerSpec.scala
@@ -19,12 +19,12 @@
 package io.renku.eventlog.subscriptions.tsmigrationrequest
 
 import Generators._
-import TypeSerializers._
 import cats.data.Kleisli
 import cats.effect.IO
 import io.renku.db.SqlStatement
-import io.renku.eventlog.InMemoryEventLogDbSpec
-import io.renku.eventlog.subscriptions.tsmigrationrequest.MigrationStatus._
+import io.renku.eventlog.MigrationStatus._
+import io.renku.eventlog.TSMigtationTypeSerializers._
+import io.renku.eventlog._
 import io.renku.events.consumers.subscriptions.{SubscriberUrl, subscriberIds, subscriberUrls}
 import io.renku.generators.CommonGraphGenerators.serviceVersions
 import io.renku.generators.Generators.Implicits._
